Connect units settings with display weather

+++ app/src/main/java/by/reshetnikov/proweather/data/apimodels/Rain.java
@@ -8,6 +8,6 @@ public class Rain {
     // Rain volume for the last 3 hours
     @SerializedName("3h")
     @Expose
-    public int rainVolume;
+    public double rainVolume;
 
 }
